[expo-cli] rename directory build-native -> eas-build
