+++ packages/expo-cli/src/commands/eas-build/__tests__/credentials-test.ts
