+++ packages/expo-cli/src/commands/eas-build/utils/git.ts
